refactor(frontend): prevent window undefined error
